# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: Домашнее задание
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: "5"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: "15"
        type: string
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: "5"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: "15"
        type: string
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        example: "null"
        type: string
      title:
        example: Погулять с собакой
        type: string
    type: object
  internal_features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        example: "5"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: "15"
        type: string
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: "5"
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: "15"
        type: string
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
...
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Просмотр списка задач с опциональной пагинацией и/или фильтрацией
        по ID автора задачи
      parameters:
      - description: Фильтрация задач по ID автора
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удалить существующую в системе задачу по её ID
      parameters:
      - description: ID удаляемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление задачи
      tags:
      - tasks
    get:
      description: Получение конкретной задачи по её ID
      parameters:
      - description: ID получаемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешной найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Обновляет информацию об уже существующей в системе задаче
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"description": "Утром в 06:30 выйти на прогулку с Бобиком"` — устанавливает новое описание для задачи
        3. **Явно передан null**: `"description": null`— очищает поле в БД (set to NULL)
        Ограничения: `title` и `completed` не могут быть выставлены как null
      parameters:
      - description: ID изменяемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
  /users:
    get:
      description: Просмотр списка пользователей с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nilchan-social_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
...
swagger: "2.0"